# Refactor index decoding