<!--
Thank you very much for contributing code or documentation to restic! Please
fill out the following questions to make it easier for us to review your
changes.

You do not need to check all the boxes below all at once, feel free to take
your time and add more commits. If you're done and ready for review, please
check the last box.
-->

What does this PR change? What problem does it solve?
-----------------------------------------------------

There were three places in the codebase that did

decode new-style index -> check error -> decode old-style index

via a system of callbacks. I've replaced all three by a single DecodeIndex that can handle both formats, saving some 20 lines of code, two public functions in internal/repository and one public error type.

There's still some repetition in the code, but I didn't see a clean way of getting rid it. Two of the call sites pass a buffer around that would have to become an argument and a fourth return value on DecodeIndex.

New-style decoding is as fast as it was before:

```
name                   old time/op  new time/op  delta
DecodeIndex-8           4.19s ± 2%   4.16s ± 1%   ~     (p=0.105 n=10+10)
DecodeIndexParallel-8   4.18s ± 1%   4.17s ± 1%   ~     (p=0.360 n=10+8)
```

Decoding old-format indices no longer requires loading and decrypting twice, so that may be faster, but there's no benchmark for that.

Was the change discussed in an issue or in the forum before?
------------------------------------------------------------

No.

Checklist
---------

- [x] I have read the [Contribution Guidelines](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#providing-patches)
- [x] I have enabled [maintainer edits for this PR](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork)
- [x] I have added tests for all changes in this PR
- [x] I have run `gofmt` on the code in all commits
- [x] All commit messages are formatted in the same style as [the other commits in the repo](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#git-commits)
- [x] I'm done, this Pull Request is ready for review